separating out stock state
