+++ dev/benchmarks/test_apps/stocks/lib/i18n/stock_strings.dart
@@ -126,6 +126,8 @@ class _StockStringsDelegate extends LocalizationsDelegate<StockStrings> {
 }
 
 StockStrings _lookupStockStrings(Locale locale) {
+  
+  
   // Lookup logic when language+country codes are specified.
   switch (locale.languageCode) {
     case 'en': {
